+++ apps/ledger-live-mobile/src/components/TransferButton.tsx
@@ -3,7 +3,7 @@ import { BoxedIcon, Flex, Tag, Text } from "@ledgerhq/native-ui";
 import { TouchableOpacity, StyleProp, ViewStyle } from "react-native";
 import type { IconType } from "@ledgerhq/native-ui/components/Icon/type";
 import { ChevronRightMedium } from "@ledgerhq/native-ui/assets/icons";
-import { track } from "../../analytics";
+import { track } from "../analytics";
 
 type Props = {
   title: string;
